+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 * Add new config option `flowman.workaround.analyze_partition` to workaround CDP 7.1 issues
 * Fix migrating Hive views to tables and vice-versa
 * Add new option "-j <n>" to allow running multiple job instances in parallel
+* Add new `uniqueKey` assertion
+* Add new `schema` assertion
 
 
 # Version 0.17.1 - 2021-06-18